+++ src/context.ts
@@ -17,7 +17,7 @@ export function createUnimport (opts: Partial<UnimportOptions>) {
   let _combinedImports: Import[] | undefined
 
   const ctx: Context = {
-    staticImports: [].concat(opts.imports).filter(Boolean),
+    staticImports: [...(opts.imports || [])].filter(Boolean),
     dynamicImports: [],
     get imports () {
       if (!_combinedImports) {
@@ -40,7 +40,7 @@ export function createUnimport (opts: Partial<UnimportOptions>) {
     // Create map
     ctx.map.clear()
     for (const _import of imports) {
-      ctx.map.set(_import.as, _import)
+      ctx.map.set(_import.as ?? _import.name, _import)
     }
 
     return imports
@@ -51,9 +51,14 @@ export function createUnimport (opts: Partial<UnimportOptions>) {
     _combinedImports = undefined
   }
 
+  function clearDynamicImports () {
+    ctx.dynamicImports.length = 0
+  }
+
   reload()
 
   return {
+    clearDynamicImports,
     modifyDynamicImports,
     getImports: () => ctx.imports,
     detectImports: (code: string) => detectImports(code, ctx),
@@ -87,7 +92,7 @@ async function detectImports (code: string, ctx: Context) {
 
   const matchedImports = Array.from(matched)
     .map(name => ctx.map.get(name))
-    .filter(i => i && !i.disabled)
+    .filter(i => i && !i.disabled) as Import[]
 
   return {
     strippedCode,